refactor into a cross account client object to handle excpetions and closing
