+++ lambda/costs-to-metrics/aws/sqs.py
@@ -4,7 +4,7 @@ import boto3
 
 
 logger = logging.getLogger()
-logger.setLevel(logging.INFO)
+logging.basicConfig(encoding='utf-8', level=logging.INFO)
 
 
 def sqs_send_message(message: str):
